requirements.txt
PyJWT
Click==7.0
Flask==1.1.1
flask-marshmallow==0.11.0
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
marshmallow==3.5.0
marshmallow-sqlalchemy==0.22.3
pkg-resources==0.0.0
PyMySQL==0.9.3
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==1.0.0
